feat: separate providers from contexts
