+++ src/components/ui/Combobox/ThemeCombo/index.tsx
@@ -3,9 +3,9 @@ import { SunMoon } from "lucide-react";
 import { FunctionComponent, useCallback, useMemo } from "react";
 import { useTranslation } from "react-i18next";
 
-import { Combobox } from "@/components/ui/Combobox/index.tsx";
+import { Combobox } from "@/components/ui/Combobox";
 import { useTheme } from "@/hooks/useTheme.tsx";
-import { Theme } from "@/providers/ThemeProvider/index.d";
+import { Theme } from "@/providers/Theme/index.d";
 
 const ThemeCombo: FunctionComponent<ThemeComboProps> = () => {
   const { t } = useTranslation(["components/theme-combo"]);
