+++ src/contexts/Font/index.d.ts
@@ -1,0 +1,4 @@
+export type FontContextType = {
+  font: File | null;
+  setFont: (font: File) => void;
+};
