+++ src/contexts/Theme/index.d.ts
@@ -1,0 +1,6 @@
+import { Theme } from "@/providers/Theme";
+
+export type ThemeProviderState = {
+  theme: Theme;
+  setTheme: (theme: Theme) => void;
+};
