+++ src/contexts/Theme/index.tsx
@@ -1,0 +1,11 @@
+import { ThemeProviderState } from "./index.d";
+import { createContext } from "react";
+
+const initialState: ThemeProviderState = {
+  theme: "system",
+  setTheme: () => null,
+};
+
+const ThemeContext = createContext<ThemeProviderState>(initialState);
+
+export default ThemeContext;
