+++ src/main.tsx
@@ -4,7 +4,7 @@ import { StrictMode } from "react";
 import { createRoot } from "react-dom/client";
 
 import App from "@/App.tsx";
-import { ThemeProvider } from "@/providers/ThemeProvider";
+import { ThemeProvider } from "@/providers/Theme";
 
 createRoot(document.getElementById("root")!).render(
   <StrictMode>
