+++ src/providers/FontProvider/index.d.ts
@@ -1,10 +1,5 @@
 import { ReactNode } from "react";
 
-export type FontContextType = {
-  font: File | null;
-  setFont: (font: File) => void;
-};
-
 export interface FontProviderProps {
   children: ReactNode;
 }
