+++ src/providers/Theme/index.d.ts
@@ -1,12 +1,9 @@
+import { ReactNode } from "react";
+
 export type Theme = "dark" | "light" | "system";
 
 export type ThemeProviderProps = {
   children: ReactNode;
   defaultTheme?: Theme;
   storageKey?: string;
-};
-
-export type ThemeProviderState = {
-  theme: Theme;
-  setTheme: (theme: Theme) => void;
 };
